+++ example/api-key.example.ts
@@ -1,0 +1,1 @@
+export const API_KEY = 'CHAT_GPT_API_KEY';
