+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "ts-chatgpt-api",
-  "version": "1.0.4",
+  "version": "1.0.5",
   "description": "",
   "main": "index.js",
   "scripts": {
