+++ src/types/chat-gpt-completion-params.ts
@@ -1,0 +1,3 @@
+export type ChatGptCompletionParams = {
+
+};
