+++ src/types/chat-gpt-error-response.ts
@@ -1,0 +1,8 @@
+export type ChatGptErrorResponse = {
+    error: {
+        message: string,
+        type: string,
+        param: any,
+        code: string
+    }
+}
