+++ src/types/chat-gpt-message.ts
@@ -1,0 +1,4 @@
+export type ChatGptMessage = {
+    role: string,
+    content: string
+}
