+++ src/types/chat-gpt-request.ts
@@ -1,0 +1,6 @@
+import {ChatGptMessage} from './chat-gpt-message';
+
+export type ChatGptRequest = {
+    model: string,
+    messages: ChatGptMessage[]
+}
