+++ src/types/chat-gpt-response-message.ts
@@ -1,0 +1,8 @@
+import {ChatGptRoles} from './chat-gpt-roles';
+
+export type ChatGptResponseMessage = {
+    role: ChatGptRoles,
+    content: string,
+    finish_reason: string,
+    index: number
+}
