+++ src/types/chat-gpt-response.ts
@@ -1,0 +1,4 @@
+import {ChatGptSuccessResponse} from './chat-gpt-success-response';
+import {ChatGptErrorResponse} from './chat-gpt-error-response';
+
+export type ChatGptResponse = ChatGptSuccessResponse | ChatGptErrorResponse;
