+++ src/types/chat-gpt-role-list.ts
@@ -1,0 +1,5 @@
+export enum ChatGptRoleList {
+    USER = 'user',
+    SYSTEM = 'system',
+    ASSISTANT = 'assistant'
+}
