+++ src/types/chat-gpt-roles.ts
@@ -1,0 +1,3 @@
+import {ChatGptRoleList} from './chat-gpt-role-list';
+
+export type ChatGptRoles = ChatGptRoleList.USER | ChatGptRoleList.ASSISTANT | ChatGptRoleList.SYSTEM;
