+++ src/types/chat-gpt-usage.ts
@@ -1,0 +1,5 @@
+export type ChatGptUsage = {
+    prompt_tokens: number,
+    completion_tokens: number,
+    total_tokens: number
+}
